+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: hleong <hleong@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/24 19:16:28 by hleong            #+#    #+#             */
-/*   Updated: 2022/11/25 17:05:40 by hleong           ###   ########.fr       */
+/*   Updated: 2022/11/28 21:24:41 by hleong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,20 +14,19 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-    int		i;
-    char	*d;
-    char	*s;
+	size_t	i;
+	char	*d;
+	char	*s;
 
-    s = (char *)src;
-    d = (char *)dst;
-    i = 0;
-
+	s = (char *)src;
+	d = (char *)dst;
+	i = 0;
 	if (d > s)
 	{
 		while (len > 0)
-		len--;	
+		len--;
 		d[i] = s[i];
 	}
 	else
-	ft_memcpy (dst, src, len);
+		ft_memcpy (dst, src, len);
 }
